Use 64-bit for average factor, if possible

+++ cpp/include/cuml/fil/treelite_importer.hpp
@@ -151,7 +151,7 @@ struct treelite_importer {
     } else {
       result = 1.0;
     }
-    return static_cast<float>(result);
+    return result;
   }
 
   auto get_bias(treelite::Model const& tl_model) { return tl_model.base_scores.AsVector(); }
